requirements.txt
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
six==1.15.0
